Refactor package architecture
