+++ src/main/java/br/unicamp/cecom/appointmentscheduler/core/features/patient/PatientEntity.java
@@ -1,20 +1,26 @@
-package br.unicamp.cecom.appointmentscheduler.core;
+package br.unicamp.cecom.appointmentscheduler.core.features.patient;
 
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.*;
+import javax.persistence.Entity;
+import javax.persistence.Id;
 import javax.validation.constraints.Max;
 import javax.validation.constraints.NotBlank;
-import java.util.UUID;
+import javax.validation.constraints.Size;
 
 @Entity
 @Data
 @NoArgsConstructor
-public class Doctor {
+@AllArgsConstructor
+@Builder
+public class PatientEntity {
 
     @Id
-    private UUID doctorId;
+    @Size(max = 11, min = 11, message = "CPF must have 11 characters")
+    private String cpf;
 
     @NotBlank(message = "Name must not be null and must contain at least one non-whitespace character")
     @Max(value = 255, message = "Name must have a maximum of 255 characters")
@@ -28,8 +34,4 @@ public class Doctor {
     @Max(value = 15, message = "Phone must have a maximum of 15 characters")
     private String phone;
 
-    @PrePersist
-    public void prePersist() {
-        this.doctorId = UUID.randomUUID();
-    }
 }
